+++ R/dependencies.R
@@ -487,29 +487,6 @@ html_dependency_stati <- function() {
 }
 
 
-html_dependency_pickr <- function() {
-  # htmlDependency(
-  #   name = "pickr",
-  #   version = "1.6.0",
-  #   src = list(href = "shinyWidgets/pickr-1.6.0", file = "assets/pickr-1.6.0"),
-  #   package = "shinyWidgets",
-  #   script = "js/pickr.min.js",
-  #   stylesheet = c(
-  #     "css/classic.min.css",
-  #     "css/monolith.min.css",
-  #     "css/nano.min.css"
-  #   ),
-  #   head = "<style>.pickr-color.disabled{cursor:not-allowed;}</style>",
-  #   all_files = FALSE
-  # )
-  htmlDependency(
-    name = "colorpickr",
-    version = "1.6.0",
-    src = c(file = "packer"),
-    package = "shinyWidgets",
-    script = "colorpickr.js"
-  )
-}